+++ src/data/duke.txt
@@ -1,2 +1,4 @@
 D , true , return book , 2019-10-15
-E , false , read book , 2019-10-15
+E , true , read book , 2019-10-15
+D , false , return book , 2019-10-15
+E , false , project meeting , 2019-10-20
