+++ src/main/java/duke/Duke.java
@@ -1,4 +1,6 @@
-import java.time.format.DateTimeParseException;
+package duke;
+
+import duke.exception.DukeException;
 
 public class Duke {
     private TaskList taskList;
