+++ src/main/java/duke/Storage.java
@@ -1,9 +1,14 @@
+package duke;
+
 import java.io.File;
 import java.io.FileNotFoundException;
 import java.io.FileWriter;
 import java.io.IOException;
 import java.util.ArrayList;
 import java.util.Scanner;
+
+import duke.exception.DukeException;
+import duke.tasks.*;
 
 public class Storage {
