+++ src/main/java/duke/exception/DukeException.java
@@ -1,5 +1,7 @@
-class DukeException  extends Exception
-{
+package duke.exception;
+
+public class DukeException  extends Exception {
+
     public DukeException (String str)
     {
         // calling the constructor of parent Exception
