+++ text-ui-test/EXPECTED.TXT
@@ -10,17 +10,17 @@ Added: borrow book
 Added: return book
 Added: project meeting
 [T][ ] borrow book
-[D][ ] return book (by: Sunday)
-[E][ ] project meeting (at: Mon 2-4pm)
+[D][ ] return book (by: Oct 15 2019)
+[E][ ] project meeting (at: Oct 20 2019)
 Nice! I've marked this task as done:
-[D][X] return book (by: Sunday)
+[D][X] return book (by: Oct 15 2019)
 Task not yet completed!
 [T][ ] borrow book
-[D][X] return book (by: Sunday)
-[E][ ] project meeting (at: Mon 2-4pm)
+[D][X] return book (by: Oct 15 2019)
+[E][ ] project meeting (at: Oct 20 2019)
 Oh no! one more thing that needs to be done:
-[D][ ] return book (by: Sunday)
+[D][ ] return book (by: Oct 15 2019)
 [T][ ] borrow book
-[D][ ] return book (by: Sunday)
-[E][ ] project meeting (at: Mon 2-4pm)
+[D][ ] return book (by: Oct 15 2019)
+[E][ ] project meeting (at: Oct 20 2019)
 Bye. See you later!
