+++ text-ui-test/input.txt
@@ -1,6 +1,6 @@
 todo borrow book
-deadline return book /by Sunday
-event project meeting /at Mon 2-4pm
+deadline return book /by 2019-10-15
+event project meeting /at 2019-10-20
 list
 mark 1
 unmark 2
